+++ TwinPushSDK/Classes/TwinPushManager.h
@@ -152,6 +152,8 @@ typedef void(^TPExternalRegisterBlock)(TPRegisterInformation* info, TPRegisterCo
 - (void)setProperty:(NSString*)name withFloatValue:(NSNumber*)value;
 - (void)setProperty:(NSString*)name withEnumValue:(NSString*)value;
 - (void)setProperty:(NSString*)name withEnumListValue:(NSArray<NSString*>*)value;
+/** Clears all the custom properties set for the device */
+- (void)clearAllProperties;
 
 #pragma mark Location
 /** Updates the current user location in TwinPush using the desired accuracy. It will initialize a CLLocationManager,